prepare interaction controller
