+++ android/lib/src/main/java/com/swmansion/gesturehandler/GestureHandlerInteractionController.java
@@ -5,4 +5,5 @@ public interface GestureHandlerInteractionController {
   boolean shouldRequireHandlerToWaitForFailure(GestureHandler handler, GestureHandler otherHandler);
   boolean shouldRecognizeSimultaneously(GestureHandler handler, GestureHandler otherHandler);
   boolean shouldHandlerBeCancelledBy(GestureHandler handler, GestureHandler otherHandler);
+  int[] getSimultaneousRelations(GestureHandler handler);
 }
